refactor(pi-ui): extract shared UI helpers
- move generic completions and filter select into `lib/ui`
- add shared input decoding and scroll range utilities
- update mode and thinking imports to use the new UI barrel
