+++ files/pi/agent/extensions/user/lib/ui/completions.ts
@@ -1,3 +1,9 @@
+/**
+ * Generic command-argument autocomplete helper, independent of any feature.
+ * Returns `null` (rather than an empty array) when nothing matches so callers
+ * can hand it straight back to pi's `getArgumentCompletions`, which treats
+ * `null` as "no completions".
+ */
 import type { AutocompleteItem } from "@earendil-works/pi-tui";
 
 export function filterCompletionsByPrefix(
