+++ files/pi/agent/extensions/user/lib/ui/index.ts
@@ -1,0 +1,6 @@
+// Generic terminal-UI primitives shared across features (not tied to any
+// single feature's domain). See the individual modules for details.
+export * from "./completions";
+export * from "./filter-select";
+export * from "./input";
+export * from "./scroll";
